requirements.txt
langdetect
tiktoken
booktype
alembic
alphamini
asyncpg
boto3
botocore
dataclasses
email-validator
fastapi
google.generativeai
librosa
midas
numpy
opencv-python
openai-whisper
pillow
protobuf==3.20.3
psycopg2-binary
pydantic
pydub
python-dotenv
python-multipart
pytesseract
requests
soundfile
sqlalchemy[asyncio]
starlette
timm
torch
torchvision
ultralytics
uvicorn[standard]
whisper
torch
transformers
aiocache==0.11.1
aioredis==1.3.1
redis==3.5.3
hf_xet
tiktoken
ujson

# RAG Chatbot Dependencies
chromadb==0.4.22
sentence-transformers==2.3.1
langchain==0.1.6
langchain-community==0.0.20
langchain-openai==0.0.5
anthropic==0.18.1
faiss-cpu==1.7.4